+++ src/app/[locale]/(main)/community/[communityId]/[postId]/page.tsx
@@ -1,0 +1,3 @@
+export default function CommunityPostDetailPage() {
+  return <div>커뮤니티 글 상세 페이지</div>;
+}
