+++ src/app/[locale]/(main)/community/[communityId]/page.tsx
@@ -1,0 +1,3 @@
+export default function CommunityPostListPage() {
+  return <div>커뮤니티 상세 페이지 (글 목록)</div>;
+}
