+++ src/app/[locale]/(main)/community/page.tsx
@@ -1,0 +1,3 @@
+export default function CommunityListPage() {
+  return <div>커뮤니티 리스트 페이지</div>;
+}
