+++ src/app/[locale]/(main)/fan-channel/[influencerId]/[postId]/page.tsx
@@ -1,0 +1,3 @@
+export default function FanChannelPostDetailPage() {
+  return <div>인플루언서 팬채널 글 상세 페이지</div>;
+}
