+++ src/app/[locale]/(main)/fan-channel/[influencerId]/page.tsx
@@ -1,0 +1,3 @@
+export default function FanChannelPostListPage() {
+  return <div>인플루언서 팬채널 상세 페이지 (글 목록)</div>;
+}
