+++ src/app/[locale]/(main)/follow/page.tsx
@@ -15,5 +15,9 @@ export async function generateMetadata({
 }
 
 export default function FollowPage() {
-  return <div className="w-full px-5">팔로우 페이지</div>;
+  return (
+    <div className="w-full px-5">
+      팔로우 페이지, 클릭 시 팬채널/인플루언서페이지/커뮤니티로 이동
+    </div>
+  );
 }
