+++ src/app/[locale]/(main)/influencer/[influencerId]/[reviewId]/page.tsx
@@ -1,0 +1,3 @@
+export default function InfluencerReviewPage() {
+  return <div>인플루언서 한줄리뷰 상세 페이지</div>;
+}
