+++ src/app/[locale]/(main)/influencer/[influencerId]/page.tsx
@@ -1,0 +1,3 @@
+export default function InfluencerPage() {
+  return <div>인플루언서 상세 페이지</div>;
+}
