+++ src/app/[locale]/(main)/my/activity-history/page.tsx
@@ -1,0 +1,3 @@
+export default function MyActivityPage() {
+  return <div>내 활동 상세내역</div>;
+}
