+++ src/app/[locale]/(main)/my/edit/_components/DeleteAccountButton.tsx
