+++ src/app/[locale]/(main)/my/edit/page.tsx
@@ -1,10 +1,11 @@
 import { Metadata } from 'next';
-
 import { getTranslations } from 'next-intl/server';
 
 import { Separator } from '@/components/ui/separator';
-import { UserAvatar, UserSettingsPanel } from '@/components/domain/user';
-import { DeleteAccountButton } from '@/components/domain/auth';
+
+import UserAvatar from '@/components/domain/user/UserAvatar';
+import UserSettingsPanel from './_components/UserSettingsPanel';
+import DeleteAccountButton from './_components/DeleteAccountButton';
 
 export async function generateMetadata({
   params: { locale },
