+++ src/app/[locale]/(main)/user/[userId]/page.tsx
@@ -1,0 +1,3 @@
+export default function UserProfilePage() {
+  return <div>타 유저 상세 페이지(활동내역)</div>;
+}
