--- src/components/domain/auth/index.tsx
@@ -1,3 +1,0 @@
-export { default as GoogleLoginButton } from './GoogleLoginButton';
-export { default as GoogleLoginHandler } from './GoogleLoginHandler';
-export { default as DeleteAccountButton } from './DeleteAccountButton';
