--- src/components/domain/influencer/index.tsx
@@ -1,1 +1,0 @@
-export { default as OnePickInfluencer } from './OnePickInfluencer';
