--- src/components/domain/menu/MenuContent.tsx
@@ -1,11 +1,0 @@
-const MenuContent = () => {
-  return (
-    <ul className="">
-      <li>인플루언서 찾기</li>
-      <li>한줄 리뷰</li>
-      <li>한줄 리뷰</li>
-      <li>고객센터</li>
-    </ul>
-  );
-};
-export default MenuContent;
