--- src/components/domain/menu/index.tsx
@@ -1,1 +1,0 @@
-export { default as MainSlideMenu } from './MainSlideMenu';
