--- src/components/domain/notification/index.tsx
@@ -1,1 +1,0 @@
-export { default as AlarmNotification } from './AlarmNotification';
