--- src/components/domain/search/index.tsx
@@ -1,1 +1,0 @@
-export { default as MainSearch } from './MainSearch';
