+++ src/components/domain/user/hooks/useUserAvatarImage.tsx
