--- src/components/domain/user/index.ts
@@ -1,3 +1,0 @@
-export { default as UserAvatar } from './profile/UserAvatar';
-export { default as MyProfileCard } from './profile/MyProfileCard';
-export { default as UserSettingsPanel } from './setting/UserSettingsPanel';
