+++ src/components/screens/index.tsx
@@ -1,0 +1,4 @@
+export { default as SignUpSuccess } from './signup/SignUpSuccess';
+export { default as MainSlideMenu } from './menu/MainSlideMenu';
+export { default as MainSearch } from './search/MainSearch';
+export { default as MainNotification } from './notification/MainNotification';
